+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.2.0] - 2020-12-30
 [0.2.0]: https://github.com/mhassan1/yarn-plugin-licenses/compare/v0.1.0...v0.2.0
 
+- Use `fs.readFilePromise` in `linker.getPackageManifest` implementations
 - Add jest test coverage
 - Bump dev dependencies
 - Migrate from tslint to eslint